service/streamlit/requirements.txt
streamlit>=1.28.0
streamlit-authenticator>=0.2.3
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
pandas>=2.0.0
minio>=7.2.0
structlog>=23.0.0
email-validator>=2.0.0
python-json-logger>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
